+++ src/Datasets/Generators/SwissRoll.php
@@ -67,18 +67,18 @@ class SwissRoll implements Generator
                                 float $depth = 21., float $noise = 0.3)
     {
         if ($scale < 0.) {
-            throw new InvalidArgumentException("Scaling factor must be greater"
+            throw new InvalidArgumentException('Scaling factor must be greater'
                 . " than 0, $scale given.");
         }
 
         if ($depth < 0) {
-            throw new InvalidArgumentException("Depth cannot be less than 0"
+            throw new InvalidArgumentException('Depth cannot be less than 0'
                 . " $depth given.");
         }
 
         if ($noise < 0.) {
-            throw new InvalidArgumentException("Noise factor cannot be less than"
-                . " 0, $noise given.");
+            throw new InvalidArgumentException('Noise factor cannot be less'
+                . " than 0, $noise given.");
         }
 
         $this->center = Vector::quick([$x, $y, $z]);